Slack interactivity, and a gate on the endpoints that carry the leads
Two things a reviewer sees before they read any code.

The alert buttons work, but Slack was stamping every alert "not configured to
handle interactive responses" -- a warning triangle on the one message whose
job is to be believed. POST /slack/interactions now acknowledges the click.
It runs no business logic, because the buttons are navigation: it verifies
Slack's HMAC in constant time, refuses anything outside a five-minute window in
either direction so a captured request cannot be replayed, and answers 200.
With no secret configured it answers 503 rather than trusting an unsigned
caller, and no refusal path can print any part of the secret or the body.

The manifest deliberately does not declare interactivity. Slack requires a
Request URL beside it, and that URL is somebody's deployment -- a hostname in a
public repo would be wrong for every reader or a pointer at our server. The
manual step is documented instead, in the manifest, the README and INSTALL.

Separately, the dashboard, /ledger and the timelines carry the companies
themselves, which on a public deployment is the pipeline the operator is
relying on being alone in having. DASHBOARD_TOKEN closes all three. It ships
empty on purpose: the evidence in this repo is only checkable while a reader
can open the deployment it came from, and /health reports which posture is
live. /health and the Pond endpoints are never gated -- the first carries
counts rather than companies, the second is how Pond health-checks the agent.

+++ scripts/make_logo.py
@@ -12,7 +12,7 @@ from PIL import Image, ImageChops, ImageDraw, ImageFilter
 
 S, SS = 512, 4
 W = S * SS
-def u(v): return int(round(v * SS))
+def u(v): return round(v * SS)
 
 INK     = (236, 243, 255)
 INK_DIM = (96, 122, 170)
